wavesurfer.js: split into multiple files
