+++ types/wavesurfer.js/src/plugin/elan.d.ts
@@ -1,21 +1,19 @@
-import { PluginDefinition, PluginParams, WaveSurferObserver, WaveSurferPlugin } from "../..";
+import * as WaveSurfer from '../../wavesurfer';
 
-export = WaveSurfer.ElanPlugin;
+export = ElanPlugin;
 
-declare namespace WaveSurfer {
-    class ElanPlugin extends WaveSurferObserver implements WaveSurferPlugin {
-        constructor(params: ElanPluginParams, ws: WaveSurfer);
-        static create(params: ElanPluginParams): PluginDefinition;
-        destroy(): void;
-        init(): void;
-    }
+declare class ElanPlugin extends WaveSurfer.Observer implements WaveSurfer.WaveSurferPlugin {
+    constructor(params: ElanPluginParams, ws: WaveSurfer.WaveSurfer);
+    static create(params: ElanPluginParams): WaveSurfer.PluginDefinition;
+    destroy(): void;
+    init(): void;
+}
 
-    interface ElanPluginParams extends PluginParams {
-        /** CSS selector or HTML element where the ELAN information should be rendered. */
-        container?: string | HTMLElement;
-        /** The location of ELAN XML data. */
-        url?: string;
-        /** If set only shows the data tiers with the TIER_ID in this map. */
-        tiers?: unknown;
-    }
+interface ElanPluginParams extends WaveSurfer.PluginParams {
+    /** CSS selector or HTML element where the ELAN information should be rendered. */
+    container?: string | HTMLElement;
+    /** The location of ELAN XML data. */
+    url?: string;
+    /** If set only shows the data tiers with the TIER_ID in this map. */
+    tiers?: unknown;
 }
