+++ types/wavesurfer.js/src/plugin/mediasession.d.ts
@@ -1,36 +1,34 @@
-import { PluginDefinition, PluginParams, WaveSurferObserver, WaveSurferPlugin } from "../..";
+import * as WaveSurfer from '../../wavesurfer';
 
-export = WaveSurfer.MediaSessionPlugin;
+export = MediaSessionPlugin;
 
-declare namespace WaveSurfer {
-    class MediaSessionPlugin extends WaveSurferObserver implements WaveSurferPlugin {
-        constructor(params: MediaSessionPluginParams, ws: WaveSurfer);
-        static create(params: MediaSessionPluginParams): PluginDefinition;
-        destroy(): void;
-        init(): void;
+declare class MediaSessionPlugin extends WaveSurfer.Observer implements WaveSurfer.WaveSurferPlugin {
+    constructor(params: MediaSessionPluginParams, ws: WaveSurfer.WaveSurfer);
+    static create(params: MediaSessionPluginParams): WaveSurfer.PluginDefinition;
+    destroy(): void;
+    init(): void;
 
-        update(): void;
+    update(): void;
 
-        readonly metadata: MediaMetadata;
-        readonly params: MediaSessionPluginParams;
-        readonly wavesurfer: WaveSurfer;
-    }
+    readonly metadata: MediaMetadata;
+    readonly params: MediaSessionPluginParams;
+    readonly wavesurfer: WaveSurfer.WaveSurfer;
+}
 
-    interface MediaSessionPluginParams extends PluginParams {
-        /** Representation of the metadata associated with a MediaSession that can be used by user agents to provide a customized user interface. */
-        metadata: MediaMetadata;
-    }
+interface MediaSessionPluginParams extends WaveSurfer.PluginParams {
+    /** Representation of the metadata associated with a MediaSession that can be used by user agents to provide a customized user interface. */
+    metadata: MediaMetadata;
+}
 
-    interface MediaMetadata {
-        title: string;
-        artist: string;
-        album: string;
-        artwork: MediaImage[];
-    }
+interface MediaMetadata {
+    title: string;
+    artist: string;
+    album: string;
+    artwork: MediaImage[];
+}
 
-    interface MediaImage {
-        src: string;
-        sizes?: string;
-        type?: string;
-    }
+interface MediaImage {
+    src: string;
+    sizes?: string;
+    type?: string;
 }
