+++ types/wavesurfer.js/src/plugin/microphone.d.ts
@@ -1,73 +1,71 @@
-import { PluginDefinition, PluginParams, WaveSurferObserver, WaveSurferPlugin } from "../..";
+import * as WaveSurfer from '../../wavesurfer';
 
-export = WaveSurfer.MicrophonePlugin;
+export = MicrophonePlugin;
 
-declare namespace WaveSurfer {
-    class MicrophonePlugin extends WaveSurferObserver implements WaveSurferPlugin {
-        constructor(params: MicrophonePluginParams, ws: WaveSurfer);
-        static create(params: MicrophonePluginParams): PluginDefinition;
-        destroy(): void;
-        init(): void;
+declare class MicrophonePlugin extends WaveSurfer.Observer implements WaveSurfer.WaveSurferPlugin {
+    constructor(params: MicrophonePluginParams, ws: WaveSurfer.WaveSurfer);
+    static create(params: MicrophonePluginParams): WaveSurfer.PluginDefinition;
+    destroy(): void;
+    init(): void;
 
-        /** Connect the media sources that feed the visualization. */
-        connect(): void;
-        /** Browser detector. */
-        detectBrowser(): Browser;
-        /** Device error callback. */
-        deviceError(code: string): void;
-        /** Disconnect the media sources that feed the visualization. */
-        disconnect(): void;
-        /** Extract browser version out of the provided user agent string. */
-        extractVersion(uastring: string, expr: string, pos: number): number;
-        /** Audio input device is ready. */
-        gotStream(stream: MediaStream): void;
-        /** Pause visualization. */
-        pause(): void;
-        /** Play visualization. */
-        play(): void;
-        /** Redraw the waveform. */
-        reloadBuffer(event: AudioProcessingEvent): void;
-        /** Allow user to select audio input device. */
-        start(): void;
-        /** Stop the device stream and remove any remaining waveform drawing from the wavesurfer canvas. */
-        stop(): void;
-        /** Stop the device and the visualization. */
-        stopDevice(): void;
-        /** Pause/resume visualization. */
-        togglePlay(): void;
+    /** Connect the media sources that feed the visualization. */
+    connect(): void;
+    /** Browser detector. */
+    detectBrowser(): Browser;
+    /** Device error callback. */
+    deviceError(code: string): void;
+    /** Disconnect the media sources that feed the visualization. */
+    disconnect(): void;
+    /** Extract browser version out of the provided user agent string. */
+    extractVersion(uastring: string, expr: string, pos: number): number;
+    /** Audio input device is ready. */
+    gotStream(stream: MediaStream): void;
+    /** Pause visualization. */
+    pause(): void;
+    /** Play visualization. */
+    play(): void;
+    /** Redraw the waveform. */
+    reloadBuffer(event: AudioProcessingEvent): void;
+    /** Allow user to select audio input device. */
+    start(): void;
+    /** Stop the device stream and remove any remaining waveform drawing from the wavesurfer canvas. */
+    stop(): void;
+    /** Stop the device and the visualization. */
+    stopDevice(): void;
+    /** Pause/resume visualization. */
+    togglePlay(): void;
 
-        readonly active: boolean;
-        readonly browser: Browser;
-        readonly bufferSize: number;
-        readonly constraints: MediaStreamConstraints;
-        readonly levelChecker: ScriptProcessorNode;
-        readonly localAudioBuffer: AudioBuffer;
-        readonly mediaStreamSource: MediaStreamAudioSourceNode;
-        readonly micContext: AudioContext;
-        readonly numberOfInputChannels: number;
-        readonly numberOfOutputChannels: number;
-        readonly params: MicrophonePluginParams;
-        readonly reloadBufferFunction: (event: AudioProcessingEvent) => void;
-        readonly stream: MediaStream;
-        readonly wavesurfer: WaveSurfer;
-    }
-
-    interface Browser {
-        browser: "firefox" | "chrome" | "edge" | "safari" | "Not a supported browser.";
-        minVersion: number | null;
-        version: number | null;
-    }
-
-    interface MicrophonePluginParams extends PluginParams {
-        /** Constraints describing the media types requested. */
-        constraints?: MediaStreamConstraints;
-        /** The buffer size in units of sample-frames (default: 4096). */
-        bufferSize?: BufferSize;
-        /** Integer specifying the number of channels for this node's input (default: 1). Values of up to 32 are supported. */
-        numberOfInputChannels?: number;
-        /** Integer specifying the number of channels for this node's output. */
-        numberOfOutputChannels?: number;
-    }
-
-    type BufferSize = 256 | 512 | 1024 | 2048 | 4096 | 8192 | 16384;
+    readonly active: boolean;
+    readonly browser: Browser;
+    readonly bufferSize: number;
+    readonly constraints: MediaStreamConstraints;
+    readonly levelChecker: ScriptProcessorNode;
+    readonly localAudioBuffer: AudioBuffer;
+    readonly mediaStreamSource: MediaStreamAudioSourceNode;
+    readonly micContext: AudioContext;
+    readonly numberOfInputChannels: number;
+    readonly numberOfOutputChannels: number;
+    readonly params: MicrophonePluginParams;
+    readonly reloadBufferFunction: (event: AudioProcessingEvent) => void;
+    readonly stream: MediaStream;
+    readonly wavesurfer: WaveSurfer.WaveSurfer;
 }
+
+interface Browser {
+    browser: "firefox" | "chrome" | "edge" | "safari" | "Not a supported browser.";
+    minVersion: number | null;
+    version: number | null;
+}
+
+interface MicrophonePluginParams extends WaveSurfer.PluginParams {
+    /** Constraints describing the media types requested. */
+    constraints?: MediaStreamConstraints;
+    /** The buffer size in units of sample-frames (default: 4096). */
+    bufferSize?: BufferSize;
+    /** Integer specifying the number of channels for this node's input (default: 1). Values of up to 32 are supported. */
+    numberOfInputChannels?: number;
+    /** Integer specifying the number of channels for this node's output. */
+    numberOfOutputChannels?: number;
+}
+
+type BufferSize = 256 | 512 | 1024 | 2048 | 4096 | 8192 | 16384;
