+++ types/wavesurfer.js/src/plugin/minimap.d.ts
@@ -1,41 +1,39 @@
-import { WaveSurferObserver, Drawer, WaveSurferPlugin, PluginDefinition, PluginParams, WaveSurferUtil } from "../..";
-import RegionsPlugin = require("./regions");
+import * as WaveSurfer from '../../wavesurfer';
+import RegionsPlugin = require('./regions');
 
-export = WaveSurfer.MinimapPlugin;
+export = MinimapPlugin;
 
-declare namespace WaveSurfer {
-    class MinimapPlugin extends WaveSurferObserver implements WaveSurferPlugin {
-        constructor(params: MinimapPluginParams, ws: WaveSurfer);
-        static create(params: MinimapPluginParams): PluginDefinition;
-        destroy(): void;
-        init(): void;
+declare class MinimapPlugin extends WaveSurfer.Observer implements WaveSurfer.WaveSurferPlugin {
+    constructor(params: MinimapPluginParams, ws: WaveSurfer.WaveSurfer);
+    static create(params: MinimapPluginParams): WaveSurfer.PluginDefinition;
+    destroy(): void;
+    init(): void;
 
-        bindMinmapEvents(): void;
-        bindWavesurferEvents(): void;
-        createElements(): void;
-        getWidth(): number;
-        moveOverviewRegion(pixels: number): void;
-        regions(): void;
-        render(): void;
-        renderRegions(): void;
+    bindMinmapEvents(): void;
+    bindWavesurferEvents(): void;
+    createElements(): void;
+    getWidth(): number;
+    moveOverviewRegion(pixels: number): void;
+    regions(): void;
+    render(): void;
+    renderRegions(): void;
 
-        readonly draggingOverview: boolean;
-        readonly drawer: Drawer;
-        readonly overviewPosition: number;
-        readonly overviewRegion: HTMLElement | null;
-        readonly overviewWidth: number | null;
-        readonly params: MinimapPluginParams;
-        readonly ratio: number;
-        readonly regionsPlugin: RegionsPlugin | null;
-        readonly renderEvent: string;
-        readonly util: WaveSurferUtil;
-        readonly waveShowedWidth: number;
-        readonly waveWidth: number;
-        readonly wavesurfer: WaveSurfer;
-    }
+    readonly draggingOverview: boolean;
+    readonly drawer: WaveSurfer.Drawer;
+    readonly overviewPosition: number;
+    readonly overviewRegion: HTMLElement | null;
+    readonly overviewWidth: number | null;
+    readonly params: MinimapPluginParams;
+    readonly ratio: number;
+    readonly regionsPlugin: RegionsPlugin | null;
+    readonly renderEvent: string;
+    readonly util: WaveSurfer.WaveSurfer['util'];
+    readonly waveShowedWidth: number;
+    readonly waveWidth: number;
+    readonly wavesurfer: WaveSurfer.WaveSurfer;
+}
 
-    interface MinimapPluginParams extends PluginParams {
-        /** CSS selector or HTML element where the map should be rendered. By default it is simply appended after the waveform. */
-        container?: string | HTMLElement | false;
-    }
+interface MinimapPluginParams extends WaveSurfer.PluginParams {
+    /** CSS selector or HTML element where the map should be rendered. By default it is simply appended after the waveform. */
+    container?: string | HTMLElement | false;
 }
